fixing camel routes

+++ weather-api-gateway/src/main/java/org/tenkichannel/weather/api/gateway/routes/openweathermap/model/Wind.java
@@ -5,7 +5,7 @@ package org.tenkichannel.weather.api.gateway.routes.openweathermap.model;
 public class Wind {
 
     private Double speed;
-    private Integer deg;
+    private Double deg;
 
     public Double getSpeed() {
         return speed;
@@ -15,11 +15,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public Integer getDeg() {
+    public Double getDeg() {
         return deg;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Double deg) {
         this.deg = deg;
     }
